extract common tests into macros
