+++ runtime/moonbase/src/tests/mod.rs
@@ -14,13 +14,33 @@
 // You should have received a copy of the GNU General Public License
 // along with Litentry.  If not, see <https://www.gnu.org/licenses/>.
 
-use crate::Runtime;
+pub mod setup {
+	use crate::Runtime;
+	#[cfg(test)]
+	runtime_common::decl_test_chain!(Runtime);
+}
 
-#[cfg(test)]
-runtime_common::decl_test_chain!(Runtime);
+mod base_call_filter {
+	use crate::{Call, Origin, Runtime};
 
-mod base_call_filter;
+	runtime_common::run_call_filter_tests!();
+}
 
-mod transaction_payment;
+mod transaction_payment {
+	use crate::{Call, Origin, Runtime, TransactionByteFee};
 
-mod xcm_parachain;
+	runtime_common::run_transaction_payment_tests!();
+}
+
+mod xcm_parachain {
+	use crate::{
+		tests::setup::{
+			Call as RelayCall, Origin as RelayOrigin, ParaA, ParaB, Relay, RelayChainRuntime,
+			TestNet,
+		},
+		xcm_config::{LocationToAccountId, UnitWeightCost},
+		Call, Origin, Runtime,
+	};
+
+	runtime_common::run_xcm_tests!();
+}
